+++ build_runner/test/build_plan/previous_build_test.dart
@@ -82,7 +82,7 @@ void main() {
 
     test('loads with no previous build state', () async {
       final previousBuild = await loadPreviousBuild();
-      expect(previousBuild.state, null);
+      expect(previousBuild.incrementalState, null);
     });
 
     Future<void> writeBuildStateAndPlan(
@@ -103,21 +103,18 @@ void main() {
       final spec = await loadSpec();
       final previousBuild = await PreviousBuild.load(spec);
       final buildState =
-          previousBuild.state?.incremental ?? IncrementalBuildState();
+          previousBuild.incrementalState ?? IncrementalBuildState();
       await writeBuildStateAndPlan(buildState, spec.buildPlanDigest);
 
       final reloadedBuild = await loadPreviousBuild();
-      expect(
-        reloadedBuild.state?.incremental.toString(),
-        buildState.toString(),
-      );
+      expect(reloadedBuild.incrementalState.toString(), buildState.toString());
     });
 
     test('discards previous build state if build phases changed', () async {
       final spec = await loadSpec();
       final previousBuild = await PreviousBuild.load(spec);
       final buildState =
-          previousBuild.state?.incremental ?? IncrementalBuildState();
+          previousBuild.incrementalState ?? IncrementalBuildState();
       await writeBuildStateAndPlan(buildState, spec.buildPlanDigest);
 
       final buildSpec2 = await BuildSpec.load(
@@ -132,7 +129,7 @@ void main() {
       );
 
       final reloadedBuild = await PreviousBuild.load(buildSpec2);
-      expect(reloadedBuild.state, null);
+      expect(reloadedBuild.incrementalState, null);
     });
 
     test('tracks lost outputs if build phases changed', () async {
@@ -155,12 +152,15 @@ void main() {
 
       await readerWriter.writeAsString(outputId, '// output');
       final step = buildStepPlan.stepForDeclaredOutput(outputId);
-      buildState.updateBuildStepResult(
-        step,
-        BuildStepResult((b) {
+      buildState.addBuildStepResult(
+        step: step,
+        result: BuildStepResult((b) {
           b.isHidden = false;
-          b.outputs[outputId] = AssetContent.digest(Digest(<int>[]));
+          b.outputs.add(outputId);
         }),
+        contents: {
+          outputId: AssetContent.string('// output', digest: Digest(<int>[])),
+        },
       );
       await writeBuildStateAndPlan(
         buildState.toIncrementalBuildState(),
@@ -179,7 +179,7 @@ void main() {
       );
 
       final reloadedBuild = await PreviousBuild.load(buildSpec2);
-      expect(reloadedBuild.state, null);
+      expect(reloadedBuild.incrementalState, null);
       expect(
         reloadedBuild.incompatibleBuildOutputsToDelete,
         contains(outputId),
@@ -190,7 +190,7 @@ void main() {
       final spec = await loadSpec();
       final previousBuild = await PreviousBuild.load(spec);
       final buildState =
-          previousBuild.state?.incremental ?? IncrementalBuildState();
+          previousBuild.incrementalState ?? IncrementalBuildState();
       await writeBuildStateAndPlan(buildState, spec.buildPlanDigest);
 
       final spec2 = await BuildSpec.load(
@@ -205,14 +205,14 @@ void main() {
       );
 
       final reloadedBuild = await PreviousBuild.load(spec2);
-      expect(reloadedBuild.state, null);
+      expect(reloadedBuild.incrementalState, null);
     });
 
     test('discards previous build state if packages changed', () async {
       final spec = await loadSpec();
       final previousBuild = await PreviousBuild.load(spec);
       final buildState =
-          previousBuild.state?.incremental ?? IncrementalBuildState();
+          previousBuild.incrementalState ?? IncrementalBuildState();
       await writeBuildStateAndPlan(buildState, spec.buildPlanDigest);
 
       final buildPackages2 = BuildPackages.singlePackageBuild('b', [
@@ -222,7 +222,7 @@ void main() {
         testingOverrides.copyWith(buildPackages: buildPackages2),
       );
 
-      expect(reloadedBuild.state, null);
+      expect(reloadedBuild.incrementalState, null);
     });
 
     test(
@@ -231,7 +231,7 @@ void main() {
         final spec = await loadSpec();
         final previousBuild = await PreviousBuild.load(spec);
         final buildState =
-            previousBuild.state?.incremental ?? IncrementalBuildState();
+            previousBuild.incrementalState ?? IncrementalBuildState();
         await writeBuildStateAndPlan(buildState, spec.buildPlanDigest);
 
         final reloadedBuild = await withEnabledExperiments(
@@ -239,13 +239,13 @@ void main() {
           ['an_experiment'],
         );
 
-        expect(reloadedBuild.state, null);
+        expect(reloadedBuild.incrementalState, null);
       },
     );
 
     test('tracks cleanBuild when build_plan.json does not exist', () async {
       final previousBuild = await loadPreviousBuild();
-      expect(previousBuild.state, isNull);
+      expect(previousBuild.incrementalState, isNull);
     });
 
     test(
@@ -254,11 +254,11 @@ void main() {
         final spec = await loadSpec();
         final previousBuild = await PreviousBuild.load(spec);
         await writeBuildStateAndPlan(
-          previousBuild.state?.incremental ?? IncrementalBuildState(),
+          previousBuild.incrementalState ?? IncrementalBuildState(),
           spec.buildPlanDigest,
         );
         final reloadedBuild = await loadPreviousBuild();
-        expect(reloadedBuild.state, isNotNull);
+        expect(reloadedBuild.incrementalState, isNotNull);
       },
     );
 
@@ -271,12 +271,12 @@ void main() {
           (b) => b.compileDigest = 'stale_digest',
         );
         await writeBuildStateAndPlan(
-          previousBuild.state?.incremental ?? IncrementalBuildState(),
+          previousBuild.incrementalState ?? IncrementalBuildState(),
           staleDigest,
         );
 
         final reloadedBuild = await loadPreviousBuild();
-        expect(reloadedBuild.state, isNull);
+        expect(reloadedBuild.incrementalState, isNull);
       },
     );
 
@@ -287,7 +287,7 @@ void main() {
         (b) => b.buildTriggersDigest = 'triggers_1',
       );
       await writeBuildStateAndPlan(
-        previousBuild.state?.incremental ?? IncrementalBuildState(),
+        previousBuild.incrementalState ?? IncrementalBuildState(),
         staleDigest,
       );
 
@@ -317,7 +317,7 @@ void main() {
       );
 
       final reloadedBuild = await PreviousBuild.load(spec2);
-      expect(reloadedBuild.state, isNotNull);
+      expect(reloadedBuild.incrementalState, isNotNull);
       expect(reloadedBuild.triggersChanged, isTrue);
     });
 
@@ -328,7 +328,7 @@ void main() {
         (b) => b.inBuildPhasesOptionsDigests[0] = 'dummy_digest_1',
       );
       await writeBuildStateAndPlan(
-        previousBuild.state?.incremental ?? IncrementalBuildState(),
+        previousBuild.incrementalState ?? IncrementalBuildState(),
         staleDigest,
       );
 
@@ -358,7 +358,7 @@ void main() {
       );
 
       final reloadedBuild = await PreviousBuild.load(spec2);
-      expect(reloadedBuild.state, isNotNull);
+      expect(reloadedBuild.incrementalState, isNotNull);
       expect(reloadedBuild.phaseOptionsChangedList[0], isTrue);
     });
 
@@ -369,12 +369,12 @@ void main() {
         (b) => b.buildPhasesDigest = 'stale_digest',
       );
       await writeBuildStateAndPlan(
-        previousBuild.state?.incremental ?? IncrementalBuildState(),
+        previousBuild.incrementalState ?? IncrementalBuildState(),
         staleDigest,
       );
 
       final reloadedBuild = await loadPreviousBuild();
-      expect(reloadedBuild.state, isNull);
+      expect(reloadedBuild.incrementalState, isNull);
     });
 
     test('tracks dartVersion when SDK version changes', () async {
@@ -384,12 +384,12 @@ void main() {
         (b) => b.dartVersion = 'stale_version',
       );
       await writeBuildStateAndPlan(
-        previousBuild.state?.incremental ?? IncrementalBuildState(),
+        previousBuild.incrementalState ?? IncrementalBuildState(),
         staleDigest,
       );
 
       final reloadedBuild = await loadPreviousBuild();
-      expect(reloadedBuild.state, isNull);
+      expect(reloadedBuild.incrementalState, isNull);
     });
 
     test('tracks enabledExperiments when experiments change', () async {
@@ -399,12 +399,12 @@ void main() {
         (b) => b.enabledExperiments.add('stale_experiment'),
       );
       await writeBuildStateAndPlan(
-        previousBuild.state?.incremental ?? IncrementalBuildState(),
+        previousBuild.incrementalState ?? IncrementalBuildState(),
         staleDigest,
       );
 
       final reloadedBuild = await loadPreviousBuild();
-      expect(reloadedBuild.state, isNull);
+      expect(reloadedBuild.incrementalState, isNull);
     });
 
     test(
@@ -416,12 +416,12 @@ void main() {
           (b) => b.packageLanguageVersions['a'] = '1.0',
         );
         await writeBuildStateAndPlan(
-          previousBuild.state?.incremental ?? IncrementalBuildState(),
+          previousBuild.incrementalState ?? IncrementalBuildState(),
           staleDigest,
         );
 
         final reloadedBuild = await loadPreviousBuild();
-        expect(reloadedBuild.state, isNull);
+        expect(reloadedBuild.incrementalState, isNull);
       },
     );
 
@@ -429,7 +429,7 @@ void main() {
       final spec = await loadSpec();
       final previousBuild = await PreviousBuild.load(spec);
       await writeBuildStateAndPlan(
-        previousBuild.state?.incremental ?? IncrementalBuildState(),
+        previousBuild.incrementalState ?? IncrementalBuildState(),
         spec.buildPlanDigest,
       );
 
@@ -442,7 +442,7 @@ void main() {
       );
 
       final reloadedBuild = await loadPreviousBuild();
-      expect(reloadedBuild.state, isNull);
+      expect(reloadedBuild.incrementalState, isNull);
     });
   });
 }