+++ build_runner/test/integration_tests/build_command_post_process_builder_test.dart
@@ -151,5 +151,45 @@ targets:
     tester.delete('root_pkg/lib/a.txt');
     await tester.run('root_pkg', 'dart run build_runner build --force-jit');
     expect(tester.read(hiddenPostOutput), null);
+
+    // Post process output outside source globs is reused on restart.
+    tester.writeFixturePackage(
+      FixturePackages.postProcessCopyBuilder(buildToCache: false),
+    );
+    tester.write('root_pkg/build.yaml', r'''
+targets:
+  $default:
+    sources:
+      - "lib/a.txt"
+''');
+    tester.write('root_pkg/lib/a.txt', 'a');
+    await tester.run('root_pkg', 'dart run build_runner build --force-jit');
+    expect(tester.read('root_pkg/lib/a.txt.post'), 'a(default dev)');
+
+    // Restart with unchanged post process output outside source globs is
+    // reused.
+    output = await tester.run(
+      'root_pkg',
+      'dart run build_runner build --force-jit',
+    );
+    expect(output, contains('wrote 0 outputs'));
+
+    // Restart with modified post process output outside source globs reruns.
+    tester.write('root_pkg/lib/a.txt.post', 'modified');
+    output = await tester.run(
+      'root_pkg',
+      'dart run build_runner build --force-jit',
+    );
+    expect(output, contains('wrote 1 output'));
+    expect(tester.read('root_pkg/lib/a.txt.post'), 'a(default dev)');
+
+    // Restart with deleted post process output outside source globs reruns.
+    tester.delete('root_pkg/lib/a.txt.post');
+    output = await tester.run(
+      'root_pkg',
+      'dart run build_runner build --force-jit',
+    );
+    expect(output, contains('wrote 1 output'));
+    expect(tester.read('root_pkg/lib/a.txt.post'), 'a(default dev)');
   });
 }